consolidate common cx functions and delete unused code
