--- src/model/import-export/cx/converter-constants.js
@@ -1,8 +1,0 @@
-export const SAVED_ASPECTS = [
-  'networkAttributes',
-  'visualEditorProperties',
-  'cyHiddenAttributes',
-  'cyTableColumn'
-];
-
-export const CX_DATA_KEY = '_cx2-data';
