--- src/model/import-export/cx/converter-utils.js
@@ -1,11 +1,0 @@
-
-
-export const isAspectKeyInArray = (aspect, aspectKeyArray) => {
-  let output = false;
-  aspectKeyArray.forEach( aspectKey => {
-      if (aspect[aspectKey]) {
-          output = true;
-      }
-  });
-  return output;
-};
