--- src/model/import-export/cx/cx-constants.js
@@ -1,26 +1,0 @@
-
-
-export const CX_VERSION = 'CXVersion';
-export const NODE= 'node';
-export const EDGE= 'edge';
-export const NETWORK= 'network';
-
-export const NODES= 'nodes';
-export const EDGES= 'edges';
-
-export const ID= 'id';
-export const X= 'x';
-export const Y= 'y';
-export const Z= 'z';
-export const V= 'v';
-
-export const AT= 'at';
-export const N= 'n';
-export const E= 'e';
-
-export const VISUAL_PROPERTIES= 'visualProperties';
-export const DEFAULT= 'default';
-
-export const STYLE= 'style';
-
-export const PO= 'po';
